Change build scripts to get the right LLVM version
